feat(v3): Phase 2 & 3 Complete - ERC-4337 Storage Optimization, DoS Protection & SDK Sync
- Optimized Operator Account storage (packed struct)
- Removed external Oracle calls in validation (PriceCache)
- Verified Security Fixes (Destination Lock, Firewall)
- Synced SDK scripts with new ABI and maxRate support

+++ contracts/test/v3/DVTSlash.t.sol
@@ -103,7 +103,7 @@ contract DVTSlashTest is Test {
         );
         
         // Verify rep loss (assuming starting rep is 100, MINOR loss is 20)
-        (,,,,,,,, uint256 reputation) = paymaster.operators(operator);
+        (,,,,, uint32 reputation,,,) = paymaster.operators(operator);
         assertEq(reputation, 80);
 
         // 3. Perform Tier 2 Slash (GToken Stake) via GTokenStaking
